--- src/App/NewGameMenu/GenerateCode/CodeActions/index.tsx
@@ -1,1 +1,0 @@
-export { CodeActions } from './CodeActions';
