--- src/App/NewGameMenu/NewGameMenu.stories.tsx
@@ -1,9 +1,0 @@
-import { NewGameMenu } from '.';
-
-export default {
-  title: 'NewGameMenu',
-  component: NewGameMenu,
-};
-
-/* eslint-disable-next-line prettier/prettier */
-export const Default = () => <NewGameMenu />;
