+++ src/app/NewGameMenu.stories.tsx
@@ -1,0 +1,8 @@
+import { NewGameMenu } from './NewGameMenu';
+
+export default {
+  title: NewGameMenu.name,
+  component: NewGameMenu,
+};
+
+export const Default = () => <NewGameMenu />;
