+++ src/app/game/guess-word-grid/FinalScoreRows.tsx
