+++ src/app/game/guess-word-grid/ScoreCells.tsx
@@ -1,4 +1,4 @@
-import type { GuessWordScore } from '../../../../domain/guess-word-service';
+import type { GuessWordScore } from '../../../domain/guess-word-service';
 
 export type ScoreCellsVariant = 'SCORE' | 'CORRECTNESS';
