+++ src/app/game/guess-word-grid/grid-template-service.ts
