+++ src/app/game/guess-word-grid/index.ts
@@ -1,0 +1,7 @@
+export { CodeLabel } from './CodeLabel';
+export { FinalScoreRows } from './FinalScoreRows';
+export { GameWordCells } from './GameWordCells';
+export { getGridTemplateAreas } from './grid-template-service';
+export { LetterCells } from './LetterCells';
+export { ScoreCells } from './ScoreCells';
+export { SubmitButton } from './SubmitButton';
