+++ src/app/index.ts
@@ -1,0 +1,2 @@
+export { Game } from './Game';
+export { NewGameMenu } from './NewGameMenu';
