+++ src/app/new-game-menu/NewGame.tsx
@@ -5,7 +5,7 @@ import Typography from '@material-ui/core/Typography';
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useInitGameState } from '../../use-cases/use-init-game-state';
-import { PlaySvg } from './GenerateCode/CodeActions/PlaySvg';
+import { PlaySvg } from './generate-code/code-actions/PlaySvg';
 
 export const NewGame = () => {
   const { GetGameWordCodeValidationRule } = useInitGameState;
