+++ src/app/new-game-menu/generate-code/code-actions/CopySvg.tsx
