+++ src/app/new-game-menu/generate-code/code-actions/PlaySvg.tsx
