+++ src/app/new-game-menu/generate-code/code-actions/index.ts
@@ -1,0 +1,3 @@
+export { CopySvg } from './CopySvg';
+export { PlaySvg } from './PlaySvg';
+export { ShareSvg } from './ShareSvg';
