+++ src/app/new-game-menu/index.ts
@@ -1,1 +1,2 @@
 export { GenerateCode } from './GenerateCode';
+export { NewGame } from './NewGame';
