+++ src/components/IndividualCharacterInputs.tsx
